Move ShareButton to separate component
